Viewpager2 delete, UI fix, Feature fix
