--- app/src/main/java/com/example/image_search/Adapter/ViewPager2Adapter.kt
@@ -1,17 +1,0 @@
-package com.example.image_search.Adapter
-
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentActivity
-import androidx.viewpager2.adapter.FragmentStateAdapter
-
-class ViewPager2Adapter(fragmentActivity: FragmentActivity) :
-    FragmentStateAdapter(fragmentActivity) {
-    private val fragments : List<Fragment> = mutableListOf()
-    override fun getItemCount(): Int {
-        return fragments.size
-    }
-
-    override fun createFragment(position: Int): Fragment {
-        return fragments[position]
-    }
-}
