+++ app/src/main/java/com/example/image_search/Favorite/FavoriteFragment.kt
@@ -1,0 +1,24 @@
+package com.example.image_search.Favorite
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.image_search.Adapter.FavoriteAdapter
+import com.example.image_search.R
+import com.example.image_search.databinding.FragmentFavoriteBinding
+
+class FavoriteFragment : Fragment() {
+
+    private lateinit var binding: FragmentFavoriteBinding
+    private lateinit var adapter: FavoriteAdapter
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentFavoriteBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+}
