+++ app/src/main/java/com/example/image_search/Favorite/FavoriteModel.kt
@@ -1,0 +1,4 @@
+package com.example.image_search.Favorite
+
+class FavoriteModel {
+}
