+++ app/src/main/java/com/example/image_search/Search/SearchFragment.kt
@@ -1,0 +1,28 @@
+package com.example.image_search.Search
+
+import android.content.Context
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import com.example.image_search.Adapter.SearchAdapter
+import com.example.image_search.BuildConfig
+import com.example.image_search.Data.SearchDTO
+import com.example.image_search.R
+import com.example.image_search.databinding.FragmentSearchBinding
+
+class SearchFragment : Fragment() {
+    private lateinit var binding: FragmentSearchBinding
+    private lateinit var adapter: SearchAdapter
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentSearchBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+}
