+++ app/src/main/java/com/example/image_search/Search/SearchModel.kt
@@ -1,0 +1,4 @@
+package com.example.image_search.Search
+
+class SearchModel {
+}
