+++ app/src/main/res/drawable/button_shape.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
+    <solid android:color="#EDC1B3"/>
+    <corners android:radius="30dp"/>
+
+</shape>
