+++ app/src/main/res/layout/fragment_search.xml
@@ -1,65 +1,40 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".SearchFragment"
-    android:id="@+id/fragment_search"
-    >
+    tools:context=".Search.SearchFragment">
 
-        <View
-            android:id="@+id/search_background"
-            android:layout_width="match_parent"
-            android:layout_height="80dp"
-            android:background="#C9C9C9"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            />
+    <ImageView
+        android:id="@+id/search_image"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:src="@mipmap/ic_launcher"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
-        <EditText
-            android:id="@+id/search_edit_text"
-            app:layout_constraintBottom_toBottomOf="@id/search_background"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            android:layout_width="240dp"
-            android:layout_height="45dp"
-            android:layout_marginStart="20dp"
-            android:background="@drawable/search_edit_text"
-            android:hint="검색어를 입력해주세요."
-            android:textSize="15sp"
-            android:padding="13dp" />
+    <TextView
+        android:id="@+id/search_title"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="10dp"
+        android:text="[Image] 네이버블로그"
+        android:textSize="13sp"
+        android:textStyle="bold"
+        app:layout_constraintEnd_toEndOf="@id/search_image"
+        app:layout_constraintStart_toStartOf="@id/search_image"
+        app:layout_constraintTop_toBottomOf="@id/search_image" />
 
-        <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/search_button_blue"
-            app:layout_constraintBottom_toBottomOf="@id/search_background"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toEndOf="@id/search_edit_text"
-            app:layout_constraintTop_toTopOf="parent"
-            android:layout_width="100dp"
-            android:layout_height="40dp"
-            android:layout_marginHorizontal="10dp"
-            android:background="@drawable/search_button"
-            android:text="검색"
-            android:textColor="@color/white"
-            android:textSize="18sp"
-            android:textStyle="bold" />
+    <TextView
+        android:id="@+id/search_date"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="2023.09.21"
+        android:textSize="13sp"
+        android:textStyle="bold"
+        app:layout_constraintEnd_toEndOf="@id/search_title"
+        app:layout_constraintStart_toStartOf="@id/search_title"
+        app:layout_constraintTop_toBottomOf="@id/search_title" />
 
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/search_background"
-            >
-
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/search_recyclerview"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                />
-        </androidx.constraintlayout.widget.ConstraintLayout>
 </androidx.constraintlayout.widget.ConstraintLayout>
