+++ app/src/main/res/layout/search_item.xml
@@ -1,51 +1,62 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/fragment_search"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    tools:context=".Search.SearchFragment">
 
-    <ImageView
-        android:id="@+id/favorite_view"
+    <View
+        android:id="@+id/search_background"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:src="@color/black"
-        android:layout_margin="20dp"
+        android:layout_height="80dp"
+        android:background="#C9C9C9"
+        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        />
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <EditText
+        android:id="@+id/search_edit_text"
+        android:layout_width="280dp"
+        android:layout_height="45dp"
+        android:layout_marginStart="20dp"
+        android:background="@drawable/search_edit_text"
+        android:hint="검색어를 입력해주세요."
+        android:padding="13dp"
+        android:textSize="15sp"
+        app:layout_constraintBottom_toBottomOf="@id/search_background"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <ImageButton
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/heart"
-        android:visibility="gone"
+        android:id="@+id/search_button"
+        android:layout_width="80dp"
+        android:layout_height="80dp"
+        android:src="@drawable/search_icon"
+        android:scaleType="fitXY"
+        android:padding="20dp"
         android:background="@android:color/transparent"
-        android:layout_marginStart="20dp" app:layout_constraintStart_toEndOf="@id/title"
-        app:layout_constraintTop_toTopOf="@id/title"
-        app:layout_constraintBottom_toBottomOf="@id/time_text"
+        app:layout_constraintStart_toEndOf="@id/search_edit_text"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="@id/search_background"
+        app:layout_constraintBottom_toBottomOf="@id/search_background"
         />
 
-    <TextView
-        android:id="@+id/title"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="[Image]네이버 블로그"
-        android:textStyle="bold"
-        android:textSize="12sp"
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
         android:layout_marginTop="10dp"
-        app:layout_constraintTop_toBottomOf="@id/favorite_view"
-        app:layout_constraintStart_toStartOf="@id/favorite_view"
-        app:layout_constraintEnd_toEndOf="@id/favorite_view"
-        />
+        android:layout_marginHorizontal="10dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/search_background">
 
-    <TextView
-        android:id="@+id/time_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="2023-09-21 20:00:00"
-        app:layout_constraintStart_toStartOf="@id/title"
-        app:layout_constraintEnd_toEndOf="@id/title"
-        app:layout_constraintTop_toBottomOf="@id/title"
-        />
-
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/search_recyclerview"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent" />
+    </androidx.constraintlayout.widget.ConstraintLayout>
 </androidx.constraintlayout.widget.ConstraintLayout>
